Time: Change old `u32` timestamps to 64-bit (#5818)
MacOSX build fix + cleanups

+++ src/guiFormSpecMenu.h
@@ -420,8 +420,8 @@ protected:
 	v2s32 m_old_pointer;  // Mouse position after previous mouse event
 	gui::IGUIStaticText *m_tooltip_element;
 
-	u32 m_tooltip_show_delay;
-	s64 m_hovered_time;
+	u64 m_tooltip_show_delay;
+	u64 m_hovered_time;
 	s32 m_old_tooltip_id;
 	std::wstring m_old_tooltip;
